Disable EndBlocker due to panic in RebalanceHook
Changes are made in the alliance/abci.go to temporarily disable the EndBlocker function. This action is taken to address the recurring issue of panic in RebalanceHook. The function will now return an empty list of ValidatorUpdate. This is a temporary fix until a solution for the RebalanceHook error is found.